[Node] Factoring `val_` out of `tag_`
It is convenient to say that all nodes have a notion of value, which can
then be used when defining what it means to show or hide a row.

A column's value is its weight (the sum of the values of the internal
nodes in that column). This allows us to define hiding and showing a row
in a uniform way.