BUGFIX: Hide rows we cannot use, initially.
The exact binary cover solver assumes that it can pick any row in the
sparse matrix as part of its loop invariant. It is possible to break
this invariant initially if we add a row with a cell whose value exceeds
the sum constraint of its column.

To fix this, we will hide such rows pro-actively (when we are creating
the cells).